BUG: Add check that each row has the same width
TODO: Give user feedback on why the parsing was faulty.

Cause of bug: If an item in a table is separated by spaces, some rows
will generate more elements and cause a crash when putting data back
into the tableview.

One possible fix is to insert empty elements where no data exists- but
this will (int most cases) not give the wanted input.